// prodis-prodiswcl/src/app/modules/prodisapi/model/ricercaProspetto.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Provincia } from './provincia';
import { StatoProspetto } from './statoProspetto';


export interface RicercaProspetto { 
    annoProspetto?: number;
    codiceFiscaleAzienda?: string;
    codiceFiscaleStudioProf?: string;
    codiceRegionale?: string;
    dataProtocolloA?: Date;
    dataProtocolloDa?: Date;
    dataRiferimentoA?: Date;
    dataRiferimentoDa?: Date;
    denominazioneAzienda?: string;
    numeroProtocollo?: number;
    provincia?: Provincia;
    statoProspettos?: Array<StatoProspetto>;
}